services/booking_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
PyJWT
python-dotenv
passlib[bcrypt]
